refactor: Move `paginate()` to `interactions.ts`.
