+++ src/interactions.ts
@@ -1,5 +1,21 @@
-import { Bot, EventHandlers, Interaction, InteractionDataOption, InteractionTypes } from 'discordeno';
+import {
+	ApplicationCommandFlags,
+	Bot,
+	ButtonComponent,
+	ButtonStyles,
+	editOriginalInteractionResponse,
+	Embed,
+	EventHandlers,
+	Interaction,
+	InteractionDataOption,
+	InteractionResponseTypes,
+	InteractionTypes,
+	MessageComponents,
+	MessageComponentTypes,
+	sendInteractionResponse,
+} from 'discordeno';
 import * as Snowflake from 'snowflake';
+import { localise, Misc } from 'logos/assets/localisations/mod.ts';
 import { addCollector, Client } from 'logos/src/client.ts';
 import configuration from 'logos/configuration.ts';
 
@@ -114,4 +130,122 @@ function parseArguments<
 	return [args as R, focused];
 }
 
-export { createInteractionCollector, parseArguments };
+/**
+ * Paginates an array of elements, allowing the user to browse between pages
+ * in an embed view.
+ */
+function paginate<T>(
+	[client, bot]: [Client, Bot],
+	interaction: Interaction,
+	{ elements, embed, view, show }: {
+		elements: T[];
+		embed: Omit<Embed, 'footer'>;
+		view: PaginationDisplayData<T>;
+		show: boolean;
+	},
+): void {
+	const data: PaginationData<T> = { elements, view, pageIndex: 0 };
+
+	const isFirst = () => data.pageIndex === 0;
+	const isLast = () => data.pageIndex === data.elements.length - 1;
+
+	const customId = createInteractionCollector([client, bot], {
+		type: InteractionTypes.MessageComponent,
+		doesNotExpire: true,
+		onCollect: (bot, selection) => {
+			if (selection.data === undefined) return;
+
+			const action = selection.data.customId!.split('|')[1]!;
+
+			switch (action) {
+				case 'PREVIOUS':
+					if (!isFirst) data.pageIndex--;
+					break;
+				case 'NEXT':
+					if (!isLast) data.pageIndex++;
+					break;
+			}
+
+			sendInteractionResponse(bot, selection.id, selection.token, {
+				type: InteractionResponseTypes.DeferredUpdateMessage,
+			});
+
+			editOriginalInteractionResponse(bot, interaction.token, {
+				embeds: [getPageEmbed(data, embed, isLast(), interaction.locale)],
+				components: generateButtons(customId, isFirst(), isLast()),
+			});
+		},
+	});
+
+	return void sendInteractionResponse(
+		bot,
+		interaction.id,
+		interaction.token,
+		{
+			type: InteractionResponseTypes.ChannelMessageWithSource,
+			data: {
+				flags: !show ? ApplicationCommandFlags.Ephemeral : undefined,
+				embeds: [getPageEmbed(data, embed, data.pageIndex === data.elements.length - 1, interaction.locale)],
+				components: generateButtons(customId, isFirst(), isLast()),
+			},
+		},
+	);
+}
+
+interface PaginationDisplayData<T> {
+	readonly title: string;
+	readonly generate: (element: T, index: number) => string;
+}
+
+interface PaginationData<T> {
+	readonly elements: T[];
+	readonly view: PaginationDisplayData<T>;
+
+	pageIndex: number;
+}
+
+function getPageEmbed<T>(data: PaginationData<T>, embed: Embed, isLast: boolean, locale: string | undefined): Embed {
+	return {
+		...embed,
+		fields: [
+			{
+				name: data.elements.length === 1
+					? data.view.title
+					: `${data.view.title} ~ Page ${data.pageIndex + 1}/${data.elements.length}`,
+				value: data.view.generate(data.elements.at(data.pageIndex)!, data.pageIndex),
+			},
+			...(embed.fields ?? []),
+		],
+		footer: isLast ? undefined : { text: localise(Misc.continuedOnNextPage, locale) },
+	};
+}
+
+function generateButtons(customId: string, isFirst: boolean, isLast: boolean): MessageComponents {
+	const buttons: ButtonComponent[] = [];
+
+	if (!isFirst) {
+		buttons.push({
+			type: MessageComponentTypes.Button,
+			customId: `${customId}|PREVIOUS`,
+			style: ButtonStyles.Secondary,
+			label: '«',
+		});
+	}
+
+	if (!isLast) {
+		buttons.push({
+			type: MessageComponentTypes.Button,
+			customId: `${customId}|NEXT`,
+			style: ButtonStyles.Secondary,
+			label: '»',
+		});
+	}
+
+	// @ts-ignore: It is guaranteed that there will be fewer or as many as 5 buttons.
+	return buttons.length === 0 ? [] : [{
+		type: MessageComponentTypes.ActionRow,
+		components: buttons,
+	}];
+}
+
+export { createInteractionCollector, paginate, parseArguments };
